// src/api/minio/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface MinioApi {
	minioPresignedUploadUrl: (request: Api.MiniosdkPresignedUploadUrlRequest) => Promise<MinioPresignedUploadUrlResponse>
}

export type MinioPresignedUploadUrlResponse = MinioPresignedUploadUrl200Response | MinioPresignedUploadUrlDefaultResponse

export interface MinioPresignedUploadUrl200Response {
	status: 200
	body: Api.MiniosdkPresignedUploadUrlResponse
	headers?: never
}

export interface MinioPresignedUploadUrlDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}